Time: 3 ms (91.08%), Space: 45.6 MB (38.14%) - LeetHub

+++ 1-two-sum/1-two-sum.java
@@ -2,7 +2,7 @@ public class Solution {
     public int[] twoSum(int[] numbers, int target) {
         HashMap<Integer,Integer> indexMap = new HashMap<Integer,Integer>();
         for(int i = 0; i < numbers.length; i++){
-            int requiredNum = (target - numbers[i]);
+            Integer requiredNum = (Integer)(target - numbers[i]);
             if(indexMap.containsKey(requiredNum)){
                 int toReturn[] = {indexMap.get(requiredNum), i};
                 return toReturn;
